+++ src/jiratui/tests/test_fulltext_search.py
@@ -17,6 +17,8 @@ from jiratui.widgets.screens.text_search import TextSearchScreen
 def app() -> JiraApp:
     config_mock = Mock(spec=ApplicationConfiguration)
     config_mock.configure_mock(
+        enable_vim_keybindings=False,
+        custom_keybindings=None,
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
         jira_api_token=SecretStr('foo'),