+++ src/jiratui/tests/test_search_results.py
@@ -21,6 +21,8 @@ from jiratui.widgets.search import (
 def app() -> JiraApp:
     config_mock = Mock(spec=ApplicationConfiguration)
     config_mock.configure_mock(
+        enable_vim_keybindings=False,
+        custom_keybindings=None,
         jira_api_base_url='foo.bar',
         jira_api_username='foo',
         jira_api_token=SecretStr('bar'),